Modify To Modularize SQL Function Files
