+++ config/passport-setup.js
@@ -1,7 +1,8 @@
 const config = require('./civic.js');
 const passport = require('passport');
 const GoogleStrategy = require('passport-google-oauth').OAuth2Strategy;
-const db = require('../db/index.js');
+const dbMAIN = require('../db/index.js');
+const db = require('../db/users.js');
 
 
 /******************************************************************************
