+++ server/index.js
@@ -5,7 +5,8 @@ const bodyParser = require('body-parser');
 const session = require('express-session');
 const cookieParser = require('cookie-parser');
 const utils = require('./utils.js')
-const db = require('../db/index.js');
+const dbMAIN = require('../db/index.js');
+const db = require('../db/users.js');
 const apiHelpers = require('../lib/apiHelper.js');
 const dataHelpers = require('../lib/dataHelpers.js')
 const apiSearch = require('../lib/apiSearch.js');
